refactor(session): split bus persistence observer
